+++ src/V1/Requirements/InvokeableRequirement.php
@@ -58,4 +58,20 @@ trait InvokeableRequirement
     {
         return $this->to($data, $fieldOrVarName);
     }
+
+    /**
+     * throws exceptions if any of our requirements are not met
+     *
+     * this is an alias of to() for readability purposes
+     *
+     * @param  mixed $data
+     *         the data to be examined by each requirement in turn
+     * @param  string $fieldOrVarName
+     *         what is the name of $data in the calling code?
+     * @return void
+     */
+    public function inspect($data, $fieldOrVarName = "value")
+    {
+        return $this->to($data, $fieldOrVarName);
+    }
 }