+++ docs/changelog.md
@@ -15,6 +15,7 @@ pageflow_next_text: Contributing
 * Tweaks and changes to improve readabilty of your code
   - added `Inspection::inspect()` as an alias for `Inspection::to()`
   - added `InvokeableAssurable::inspect()`
+  - added `InvokeableRequirement::inspect()`
 
 * Added support for applying assurances and requirements to all elements in a list
   - added `ListInspection` interface